prototype/requirements.txt
aniso8601==8.0.0
click==7.1.2
cx-Oracle==7.3.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.1
